Update q1_memory function
Remove unnecessary intermediate variables and add some  statements.

Update dict to defaultdict for a shorter code after seeing it doesn't affect the memory efficiency of the function

Remove the usage of the Garbage Collector. It was not doing nothing impressive for the memory usage and it was adding delay to the code. Python handles garbage collection in a great way by itself